Use Passport to Add Cookies and Sessions (Level 5)
Add level 5 security/authentication by using Passport.js to implement cookies that the browser stores which hold the user login credentials for that session. This allows the website to 'remember' when a user is already logged in so that when they browse around, or access a part of the website that requires authentication, they don't need to keep logging in.